+++ src/components/DownloadComponents/SchoolDetails/SchoolDetails.tsx
@@ -6,6 +6,7 @@ import SchoolPicker from "@/components/SchoolPicker";
 import useSchoolPicker from "@/components/SchoolPicker/useSchoolPicker";
 import { Heading } from "@/components/Typography";
 import Checkbox from "@/components/Checkbox";
+import FieldError from "@/components/FormFields/FieldError";
 
 export type SchoolDetailsProps = {
   setSchool: (value: string, name?: string) => void;
@@ -74,18 +75,18 @@ const SchoolDetails: FC<SchoolDetailsProps> = ({
   };
   return (
     <>
-      <Heading tag="h2" $font={"heading-5"} $mb={16}>
+      <Heading tag="h2" $font={["heading-6", "heading-5"]} $mb={[24, 32]}>
         Your details
       </Heading>
-      <Heading tag="h3" $font={"heading-7"} $mt={0} $mb={24}>
-        Find your school in the field below (required)
-      </Heading>
+      {errors?.school && (
+        <FieldError id="school-error">{errors?.school?.message}</FieldError>
+      )}
       <SchoolPicker
         hasError={errors?.school !== undefined}
         schoolPickerInputValue={schoolPickerInputValue}
         setSchoolPickerInputValue={onSchoolPickerInputChange}
         schools={schools}
-        label={"Name of school"}
+        label={"School"}
         setSelectedSchool={setSelectedSchool}
         required={true}
       />