+++ src/pages/about-us/work-with-us.tsx
@@ -1,23 +1,23 @@
 import { NextPage, GetStaticProps, GetStaticPropsResult } from "next";
-import { PortableText } from "@portabletext/react";
 import { Fragment } from "react";
 
-import CMSClient from "../../node-lib/cms";
-import { AboutWorkWithUsPage } from "../../common-lib/cms-types";
-import Layout from "../../components/Layout";
-import MaxWidth from "../../components/MaxWidth/MaxWidth";
-import Card from "../../components/Card";
-import AboutContactCard from "../../components/AboutContactCard";
-import { Heading } from "../../components/Typography";
-import Typography from "../../components/Typography/Typography";
-import ButtonAsLink from "../../components/Button/ButtonAsLink";
-import Flex from "../../components/Flex";
-import Grid, { GridArea } from "../../components/Grid";
-import AboutIntroCard from "../../components/AboutIntoCard/AboutIntroCard";
-import BrushBorders from "../../components/SpriteSheet/BrushSvgs/BrushBorders";
-import AboutUsSummaryCard from "../../components/pages/AboutUs/AboutUsSummaryCard";
-import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
-import getPageProps from "../../node-lib/getPageProps";
+import CMSClient from "@/node-lib/cms";
+import { AboutWorkWithUsPage } from "@/common-lib/cms-types";
+import Layout from "@/components/Layout";
+import MaxWidth from "@/components/MaxWidth/MaxWidth";
+import Card from "@/components/Card";
+import AboutContactCard from "@/components/AboutContactCard";
+import { Heading } from "@/components/Typography";
+import Typography from "@/components/Typography/Typography";
+import ButtonAsLink from "@/components/Button/ButtonAsLink";
+import Flex from "@/components/Flex";
+import Grid, { GridArea } from "@/components/Grid";
+import AboutIntroCard from "@/components/AboutIntoCard/AboutIntroCard";
+import BrushBorders from "@/components/SpriteSheet/BrushSvgs/BrushBorders";
+import AboutUsSummaryCard from "@/components/pages/AboutUs/AboutUsSummaryCard";
+import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
+import getPageProps from "@/node-lib/getPageProps";
+import { PortableTextWithDefaults } from "@/components/PortableText";
 
 export type AboutPageProps = {
   pageData: AboutWorkWithUsPage;
@@ -56,7 +56,7 @@ const AboutUsBoard: NextPage<AboutPageProps> = ({ pageData }) => {
                     {card.title}
                   </Heading>
                   <Typography $mb={32} $font={["body-2", "body-1"]}>
-                    <PortableText value={card.bodyPortableText} />
+                    <PortableTextWithDefaults value={card.bodyPortableText} />
                   </Typography>
                   {card.cta?.linkType == "external" && (
                     <Flex>