+++ src/pages/legal/[policyPageSlug].tsx
@@ -4,22 +4,23 @@ import {
   GetStaticPropsResult,
   NextPage,
 } from "next";
-import { PortableText, PortableTextComponents } from "@portabletext/react";
+import { PortableTextComponents } from "@portabletext/react";
 
-import CMSClient from "../../node-lib/cms";
-import { PolicyPage } from "../../common-lib/cms-types";
+import CMSClient from "@/node-lib/cms";
+import { PolicyPage } from "@/common-lib/cms-types";
 import {
   getFallbackBlockingConfig,
   shouldSkipInitialBuild,
-} from "../../node-lib/isr";
-import Flex from "../../components/Flex";
-import Grid, { GridArea } from "../../components/Grid";
-import Layout from "../../components/Layout";
-import MaxWidth from "../../components/MaxWidth/MaxWidth";
-import Typography, { Heading, P } from "../../components/Typography";
-import { BasePortableTextProvider } from "../../components/PortableText";
-import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
-import getPageProps from "../../node-lib/getPageProps";
+} from "@/node-lib/isr";
+import Flex from "@/components/Flex";
+import Grid, { GridArea } from "@/components/Grid";
+import Layout from "@/components/Layout";
+import MaxWidth from "@/components/MaxWidth/MaxWidth";
+import Typography, { Heading, P } from "@/components/Typography";
+import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
+import getPageProps from "@/node-lib/getPageProps";
+import { PortableTextWithDefaults } from "@/components/PortableText";
+import OakLink from "@/components/OakLink";
 
 type SerializedPolicyPage = Omit<PolicyPage, "lastUpdatedAt"> & {
   lastUpdatedAt: string;
@@ -67,6 +68,24 @@ const customPolicyComponent: PortableTextComponents = {
       </P>
     ),
   },
+  marks: {
+    link: ({ children, value }) => {
+      let ariaLabel = "";
+      if (Array.isArray(children)) {
+        ariaLabel = children[0];
+      }
+      return (
+        <OakLink
+          href={value?.href}
+          aria-label={ariaLabel}
+          $textDecoration={"underline"}
+          page={null}
+        >
+          {children}
+        </OakLink>
+      );
+    },
+  },
 };
 
 const Policies: NextPage<PolicyPageProps> = ({ policy }) => {
@@ -98,12 +117,11 @@ const Policies: NextPage<PolicyPageProps> = ({ policy }) => {
               </time>
             </P>
             <Typography>
-              <BasePortableTextProvider>
-                <PortableText
-                  value={policy.bodyPortableText}
-                  components={customPolicyComponent}
-                />
-              </BasePortableTextProvider>
+              <PortableTextWithDefaults
+                value={policy.bodyPortableText}
+                components={customPolicyComponent}
+                withoutDefaultComponents
+              />
             </Typography>
           </GridArea>
         </Grid>